create_mapping, embedding tests
